requirements.txt
asgiref==3.5.2; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.8"
certifi==2022.6.15; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.1" and python_full_version < "4.0.0"
cffi==1.15.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
charset-normalizer==2.1.0; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.1" and python_full_version < "4.0.0"
coreapi==2.3.3; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
coreschema==0.0.4; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
cryptography==37.0.4; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
defusedxml==0.7.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
django-templated-mail==1.1.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
django==4.0.6; python_version >= "3.8"
djangorestframework-simplejwt==4.8.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.7"
djangorestframework==3.13.1; python_version >= "3.6"
djoser==2.1.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
idna==3.3; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.1" and python_full_version < "4.0.0"
itypes==1.2.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
jinja2==3.1.2; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.7"
markupsafe==2.1.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.7"
oauthlib==3.2.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
pycparser==2.21; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
pyjwt==2.4.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.7"
python3-openid==3.2.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
pytz==2022.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.7"
requests-oauthlib==1.3.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
requests==2.28.1; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.1" and python_full_version < "4.0.0"
six==1.16.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
social-auth-app-django==4.0.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
social-auth-core==4.3.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
sqlparse==0.4.2; python_version >= "3.8"
tzdata==2022.1; sys_platform == "win32" and python_version >= "3.8"
uritemplate==4.1.1; python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version >= "3.6"
urllib3==1.26.10; python_full_version >= "3.6.1" and python_version < "4" and python_version >= "3.7" and python_full_version < "4.0.0"
